refactor(backend): isolate gateway status probing
